requirements.txt
streamlit>=1.20.0
openai-whisper
torch>=1.12.0
sounddevice
soundfile
numpy
gTTS
pygame
opencv-python
requests